Java Version from Maven POM
Detects the java version from a pom.xml file. Uses standard maven processing libraries
to infer a version of the JVM that could be used in a buildpack from a typical pom. If
no version is specified, none is returned to the end user.

Usage: java -jar buildpack-jvm-utils-${VERSION}.jar javaversion -pom /path/to/pom.xml